V_2/requirements.txt
asgiref==3.3.4
autopep8==1.5.6
cffi==1.14.5
Django==3.1.7
django-cors-headers==3.7.0
django-simple-history==2.12.0
djangorestframework==3.12.4
pathlib==1.0.1
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
